Create Official Connect Four (advanced)
This was more of an exercise in getting chatgpt to code this correctly for me. You have to get really specific and have some knowledge of coding to correct the mistakes chatgpt makes. I realized you have to go one problem at a time and ask it to fix things along the way. making this took me 3 different sessions and over 45-50 total prompts. I made 36 versions all together.